GB: heavy rewrite of the iRDM (3.10) section
Summation indices were misleading, the normal case was written (in the formula) to be diagonal, which is not for Norb > 1. The listing were correct, so it was just a matter of notational mess. I also added a formal proof to why the normal iRDM is block-diagonal with respect to impurity Nup and Ndw sectors.
It should read pretty smooth now.

I also edited many sub- and super-scripts, adding the usual \mathrm{}, when they are words (bath, imp, int, e-ph, ...)

Finally, I did some light editing on the Attractive Hubbard example, just because I found some interpretation statements a bit too opinionated. More in general, given the scope of the paper, I'd say we don't need to delve that much into physics, do we? I thought it should really just be about showcasing what can be computed and how to do so with each API. Like, we could totally get rid of the BCS results, as they are not retrieved with EDIpack2.