GB pass
- fix a few typos and minor things
- move Holstein example after the Attractive Hubbard example (I feel like clustering the superc example is a nice idea).

Currently deciding if I will add a Edipack2ineq example (as we never give any example for using that module). Probably some nonsu2 stuff with the KMH model, so it fits well just after the BHZ excitons. We might want also a Julia API example, but I'm really not in the mood to install the Julia runtime and start tinkering with it.